01_projects/code/python-playwright/Zainar Assessment/playwright-framework/requirements.txt
pytest==7.4.3
playwright==1.57.0
pytest-playwright==0.7.2
pytest-html==4.1.1
pyyaml==6.0.1
python-dotenv==1.0.0
